--- modulo3/projeto-whatslab/whatslab/src/Components/Mensagens/Mensagens.js
@@ -1,23 +1,0 @@
-import styled from "styled-components";
-
-export const Tela = styled.div `
-    display: flex;
-    width: 60%;
-    height: 530px;
-    background-color: white;
-    align-items: flex-end;
-    margin: 0 auto;
-    border-radius: 20px 0px;
-` 
-
-export const Texto = styled.div `
-    display: flex;
-    flex-direction: column ;
-    align-items:flex-start ;
-    width: 150px;
-    height: 100px;
-    background-color: white;
-    border-radius: 20px ;
-    margin: 10px;
-    
-`
